check_integrity: Simple system integrity checker
Why another one of them checkers? Because on some servers, you don't have
a second system that can take care of verifying the integrity of the
integrity checker itself. Which makes systems like Tripwire that sign
their database to make sure that attackers can't tamper with the data an
useless overhead. And a pain in the ass, because you still have to sign
off every change in the system, knowing very well that an attacker could
simply patch the tool to ignore his malicious changes and you wouldn't
even notice. Well.  Remove the signing, and you're left with essentially
debsums for the whole system. Which is way too simple to risk using a
large and potentially buggy tool, when you can also write a few 100 LOC
to achieve the same goal on your own.